refactor bot code into its own file per service
